Split widget provider into focused files, add first JVM unit tests
HomeWidgetProvider.kt had grown to ~800 lines mixing rendering,
theming, formatting helpers and SharedPreferences persistence.
Extracted three files with no behavior change (all extracted symbols
went from private-in-file to internal so HomeWidgetProvider.kt can
still call them — same-package top-level declarations need no import):

- WidgetTheme.kt: design tokens (spacing/radii/text sizes), the
  ColorProvider(day, night) factory, and the Material You dynamic-
  color resolution (WidgetAccentColors / widgetAccentColors()).
- WidgetFormatting.kt: the pure, Android-framework-light helpers
  (formatWidgetAmount, widgetColor, calculateBitmapSampleSize) —
  deliberately kept separate from anything touching Uri/Context so
  they're plain-JVM-testable without Robolectric.
- HomeWidgetPrefs.kt: the per-widget-id SharedPreferences config
  store, already fully self-contained.

Also added the app module's first JVM unit test source set
(android/app/src/test/, new testImplementation("junit:junit:4.13.2")
in build.gradle.kts — there was no test infrastructure at all before
this) with WidgetFormattingTest covering the three extracted helpers,
including a regression test for the Color(Int) vs Color(ULong) bug
fixed earlier this session.
