+++ CHANGELOG.md
@@ -56,6 +56,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed a group's **Share this group** toggle showing as ON and remaining tappable even when neither Wi-Fi nor Bluetooth sync is turned on app-wide — sharing can't actually be active without a channel, so the switch now only displays on and reacts to taps while at least one is enabled, and the "Manage device pairing" section beneath it (QR/Bluetooth/paired devices) stays hidden along with it
 
 ### Changed
+- **Split the Android widget's native Kotlin implementation out of one ~800-line file** (`HomeWidgetProvider.kt`) into `WidgetTheme.kt` (design tokens, dynamic-color logic), `WidgetFormatting.kt` (pure amount/color/bitmap-sizing helpers) and `HomeWidgetPrefs.kt` (per-widget config persistence), leaving `HomeWidgetProvider.kt` focused on what the widget actually renders. Also added the module's first JVM unit tests (`android/app/src/test/`, new `testImplementation("junit:junit")` dependency) covering the now-extracted pure helpers — including a regression test for the `Color(Int)` vs `Color(ULong)` bug fixed earlier this session
 - **Consolidated the native (non-Flutter-engine) SQLite reads used by the Android widget and App Functions behind two SQL views** instead of five separate hand-rolled queries in `AppFunctionStorageReader.kt`: `group_summary_view` (active groups, "archived = 0" filter, today/week/total spend per group) now backs `getActiveGroupsSqlite`, `getTotalBalanceSqlite` and `getWidgetTotalsSqlite`, and a new `expense_details_view` (expenses pre-joined with category/participant names) backs `getRecentExpensesSqlite`. Same behavior, but the "which groups are active" filter and the expenses JOIN each exist in exactly one place instead of being repeated per query
 - The Android home-screen widget's native default now varies by flavor: enabled on `dev` and `staging`, disabled on `prod` (`android/app/build.gradle.kts` now sets `enable_android_widget` per product flavor instead of one global default). CI's `staging` builds pass `ENABLE_ANDROID_WIDGET=true` and `prod` builds pass `=false` to match; an explicit `enableAndroidWidget` Gradle property or `ENABLE_ANDROID_WIDGET` env var still overrides the default for every flavor
 - Sync History is now a dedicated full page (Settings → Sync → Sync History) instead of a modal bottom sheet, and is also reachable per-group from that group's own Sync sub-page, scoped to sync events that touched that group specifically
