+++ android/app/src/main/kotlin/org/app/caravella/WidgetTheme.kt
@@ -1,0 +1,143 @@
+package io.caravella.egm
+
+import android.content.Context
+import android.content.res.Configuration
+import android.os.Build
+import androidx.compose.material3.dynamicDarkColorScheme
+import androidx.compose.material3.dynamicLightColorScheme
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.glance.unit.ColorProvider
+
+// Design tokens (spacing, radii, text sizes) for the widget's Compose/Glance
+// layouts, kept separate from HomeWidgetProvider.kt's rendering logic.
+
+internal val WidgetInnerPadding = 12.dp
+internal val WidgetCompactPadding = 6.dp
+internal val WidgetSectionSpacing = 10.dp
+internal val WidgetMinimalSpacing = 2.dp
+internal val WidgetOuterRadius = 24.dp
+internal val WidgetBodyTextSize = 15.sp
+internal val WidgetLabelTextSize = 12.sp
+internal val WidgetCompactValueTextSize = 16.sp
+internal val WidgetGroupTotalValueTextSize = 22.sp
+internal val WidgetTodayPillRadius = 16.dp
+internal val WidgetTodayPillHorizontalPadding = 8.dp
+internal val WidgetTodayPillVerticalPadding = 2.dp
+internal val WidgetCtaButtonRadius = 16.dp
+internal val WidgetCtaButtonPadding = 14.dp
+internal val WidgetCtaButtonTextSize = 24.sp
+
+// Smaller CTA chip variant used by the 4x1 layout, whose ~51dp height can't fit
+// the full-size CTA button used by the taller layouts.
+internal val WidgetCtaButtonSmallRadius = 12.dp
+internal val WidgetCtaButtonSmallPadding = 6.dp
+internal val WidgetCtaButtonSmallTextSize = 16.sp
+internal val WidgetCtaButtonSmallReservedWidth = 40.dp
+
+/**
+ * Creates a [ColorProvider] that resolves to [day] in light mode and [night] in dark mode.
+ * Replaces the removed `androidx.glance.color.ColorProvider(day, night)` factory.
+ */
+internal fun ColorProvider(day: Color, night: Color): ColorProvider {
+    return DayNightColorProvider(day, night)
+}
+
+private data class DayNightColorProvider(val day: Color, val night: Color) : ColorProvider {
+    override fun getColor(context: Context): Color {
+        val nightMode = context.resources.configuration.uiMode and
+            Configuration.UI_MODE_NIGHT_MASK
+        return if (nightMode == Configuration.UI_MODE_NIGHT_YES) night else day
+    }
+}
+
+// Default widget container surface when group-based background is disabled.
+// Glass-like overlay above image/color backgrounds for better text readability.
+// Uses consistent ~80% opacity in both themes for predictable legibility.
+// This is layered on top of custom group image/color backgrounds.
+/**
+ * Returns the content overlay surface color with alpha based on the transparency level.
+ * transparency=0 → full overlay (0xCC alpha ≈ 80%), transparency=100 → no overlay (fully transparent).
+ */
+internal fun contentOverlaySurface(transparency: Int): ColorProvider {
+    // Base alpha is 0xCC (204) at transparency=0; linearly decreases to 0x00 at transparency=100.
+    val clampedTransparency = transparency.coerceIn(0, 100)
+    val alpha = ((100 - clampedTransparency) * 0xCC / 100.0).toInt()
+    val alphaHex = alpha.toLong()
+    return ColorProvider(
+        Color((alphaHex shl 24) or 0xFFFFFF),  // Light mode (white overlay)
+        Color((alphaHex shl 24) or 0x000000),  // Dark mode (black overlay)
+    )
+}
+
+internal val EmphasisTextColor = ColorProvider(
+    Color(0xFF1D1A24), // Light mode
+    Color(0xFFF4EEFF), // Dark mode
+)
+
+internal val SecondaryTextColor = ColorProvider(
+    Color(0xFF5F5A68), // Light mode
+    Color(0xFFC8C2D2), // Dark mode
+)
+
+// Static palette used when Material You dynamic color isn't available (API <31,
+// or the system theme can't be resolved for some reason) — same purple accent
+// the widget always used before dynamic color support was added.
+private val FallbackTodayPillSurface = ColorProvider(
+    Color(0x80E1D8F8), // Light mode – 50% opacity
+    Color(0x804D3B73), // Dark mode – 50% opacity
+)
+
+private val FallbackTodayPillTextColor = ColorProvider(
+    Color(0xFF2E1B52), // Light mode
+    Color(0xFFF3ECFF), // Dark mode
+)
+
+private val FallbackCtaButtonSurface = ColorProvider(
+    Color(0xFF6750A4), // Light mode (Material primary)
+    Color(0xFFD0BCFF), // Dark mode
+)
+
+private val FallbackCtaButtonTextColor = ColorProvider(
+    Color(0xFFFFFFFF), // Light mode
+    Color(0xFF381E72), // Dark mode
+)
+
+/** The widget's two branded accent surfaces (CTA button, "today" pill), themed together. */
+internal data class WidgetAccentColors(
+    val ctaSurface: ColorProvider,
+    val ctaText: ColorProvider,
+    val pillSurface: ColorProvider,
+    val pillText: ColorProvider,
+)
+
+private val FallbackWidgetAccentColors = WidgetAccentColors(
+    ctaSurface = FallbackCtaButtonSurface,
+    ctaText = FallbackCtaButtonTextColor,
+    pillSurface = FallbackTodayPillSurface,
+    pillText = FallbackTodayPillTextColor,
+)
+
+/**
+ * Resolves the widget's accent colors from the device's Material You wallpaper
+ * palette (Android 12+), so the CTA button and "today" pill follow the user's
+ * chosen system theme instead of a fixed purple — matching the rest of the app,
+ * which already supports dynamic color. Falls back to the static purple palette
+ * on older Android versions, or if the system palette can't be resolved.
+ */
+internal fun widgetAccentColors(context: Context): WidgetAccentColors {
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.S) return FallbackWidgetAccentColors
+    return try {
+        val light = dynamicLightColorScheme(context)
+        val dark = dynamicDarkColorScheme(context)
+        WidgetAccentColors(
+            ctaSurface = ColorProvider(day = light.primary, night = dark.primary),
+            ctaText = ColorProvider(day = light.onPrimary, night = dark.onPrimary),
+            pillSurface = ColorProvider(day = light.primaryContainer, night = dark.primaryContainer),
+            pillText = ColorProvider(day = light.onPrimaryContainer, night = dark.onPrimaryContainer),
+        )
+    } catch (_: Exception) {
+        FallbackWidgetAccentColors
+    }
+}
